+++ src/main/java/com/travelservice/domain/user/controller/UserController.java
@@ -7,8 +7,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.travelservice.domain.user.dto.LoginRequestDto;
-import com.travelservice.domain.user.dto.LoginResponseDto;
 import com.travelservice.domain.user.dto.UserRegistrationRequestDto;
 import com.travelservice.domain.user.entity.User;
 import com.travelservice.domain.user.service.UserService;
@@ -25,12 +23,4 @@ public class UserController {
 		User user = userService.registerMember(requestDto);
 		return ResponseEntity.ok(user);
 	}
-
-	//로그인
-	@PostMapping("/login")
-	public ResponseEntity<LoginResponseDto> login(@RequestBody LoginRequestDto requestDto) {
-		LoginResponseDto response = userService.login(requestDto);
-		return ResponseEntity.ok(response);
-	}
-
 }
